pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "tradebot"
version = "0.1.0"
description = "Kalshi near-resolution mispricing trading bot"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "cryptography>=42.0.0",
    "apscheduler>=3.10.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "httpx"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
